Move the prawn monkey-patch into an initializer.
It was in separate controllers, and this was causing issues when performing background jobs, because it
would use a different instance of the code depending on which controller was loaded last.
