+++ config/initializers/prawn_labels.rb
@@ -127,3 +127,24 @@ Prawn::Labels.types = {
     "right_margin" => 21.6
   }
 }
+
+# Monkey-batch Prawn-labels so that I can adjust the expected line-length requirement.
+module Prawn
+  class Labels
+    def shrink_text(record)
+      linecount = (split_lines = record.split("\n")).length
+
+      # 15 is estimated max character length per line.
+      split_lines.each { |line| linecount += line.length / 13 } # Total hack to make square labels work.
+
+      # -10 accounts for the overflow margins
+      rowheight = @document.grid.row_height - 10
+
+      if linecount <= rowheight / 12.floor
+        @document.font_size = 12
+      else
+        @document.font_size = rowheight / (linecount + 1)
+      end
+    end
+  end
+end
